Separated generators from the main library
