+++ language-plutus-core/generators/Language/PlutusCore/Generators/Interesting.hs
@@ -1,0 +1,1 @@
+module Language.PlutusCore.Generators.Interesting where
