--- language-plutus-core/src/Language/PlutusCore/TestSupport.hs
@@ -1,10 +1,0 @@
--- | Reexports from modules from the @TestSupport@ folder.
-
-module Language.PlutusCore.TestSupport
-    ( module Export
-    ) where
-
-import           Language.PlutusCore.TestSupport.Denotation      as Export
-import           Language.PlutusCore.TestSupport.Generator       as Export
-import           Language.PlutusCore.TestSupport.TypedBuiltinGen as Export
-import           Language.PlutusCore.TestSupport.Utils           as Export
