--- plutus-core-vm/README.md
@@ -1,15 +1,0 @@
-# Plutus Core VM Library
-
-## Specification
-
-### Exported functionality: `Language.PlutusCore.VM`
-
-
-### Testing
-
-
-### Documentation
-
-
-
-## Design
